Move zustand provider into context directory
